Unify AudioDecoder and MidiOut. They use now a common base class and only implement the different behaviors.
