+++ CMakeLists.txt
@@ -35,15 +35,15 @@ add_library(${PROJECT_NAME} STATIC
 	src/audio.cpp
 	src/audio_decoder.cpp
 	src/audio_decoder.h
+	src/audio_decoder_base.cpp
+	src/audio_decoder_base.h
+	src/audio_decoder_midi.cpp
+	src/audio_decoder_midi.h
 	src/audio_generic.cpp
 	src/audio_generic.h
 	src/audio.h
 	src/audio_midi.cpp
 	src/audio_midi.h
-	src/audio_midiout.cpp
-	src/audio_midiout.h
-	src/audio_midiout_device.cpp
-	src/audio_midiout_device.h
 	src/audio_resampler.cpp
 	src/audio_resampler.h
 	src/audio_sdl.cpp
@@ -81,8 +81,6 @@ add_library(${PROJECT_NAME} STATIC
 	src/decoder_fluidsynth.h
 	src/decoder_libsndfile.cpp
 	src/decoder_libsndfile.h
-	src/decoder_midigeneric.cpp
-	src/decoder_midigeneric.h
 	src/decoder_mpg123.cpp
 	src/decoder_mpg123.h
 	src/decoder_oggvorbis.cpp
